// MedManage.Modern/client/medmanage-angular/src/app/features/members/models/member.models.ts
export interface MemberChronicIllnessDto {
  id: number;
  memberId: number;
  chronicIllnessId: number;
  chronicIllnessName: string;
  dateFrom: string | null;
  dateTo: string | null;
  comments: string | null;
}

export interface CreateMemberChronicIllnessRequest {
  chronicIllnessId: number;
  dateFrom?: string | null;
  dateTo?: string | null;
  comments?: string | null;
}